+++ Sources/KeyboardKit/Callouts/InputCalloutContext.swift
@@ -9,15 +9,14 @@
 import SwiftUI
 
 /**
- This context can be used to handle input callouts that show
- the currently typed character.
+ This context can be used to handle callouts that show a big
+ version of the currently typed character.
  
  You can inherit this class and override any open properties
  and functions to customize the standard behavior.
  
- `KeyboardKit` will automatically create an instance of this
- class and bind it to the `KeyboardInputViewController`. The
- default instance will only be enabled for iPhone.
+ KeyboardKit automatically creates an instance of this class
+ and binds it to the ``KeyboardInputViewController``.
  */
 open class InputCalloutContext: ObservableObject {