+++ Sources/KeyboardKit/Layout/Providers/iPadKeyboardLayoutProvider.swift
@@ -258,9 +258,9 @@ struct iPadKeyboardLayoutProvider_Previews: PreviewProvider {
             layout: layout(for: locale).keyboardLayout(for: context),
             appearance: StandardKeyboardAppearance(context: context),
             actionHandler: .preview,
-            context: .preview,
-            inputContext: .preview,
-            secondaryInputContext: .preview,
+            keyboardContext: .preview,
+            actionCalloutContext: .preview,
+            inputCalloutContext: .preview,
             width: context.previewWidth)
             .background(previewImage)
             .background(Color.gray)