+++ Sources/KeyboardKit/_Deprecated/KeyboardAppearance+Deprecated.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public extension KeyboardAppearance {
+    
+    func secondaryInputCalloutStyle() -> ActionCalloutStyle {
+        actionCalloutStyle()
+    }
+}