+++ Sources/KeyboardKit/_Deprecated/SecondaryInputCalloutContext.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "ActionCalloutContext")
+public typealias SecondaryInputCalloutContext = ActionCalloutContext