+++ Sources/KeyboardKit/_Deprecated/SecondaryInputCalloutStyle.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "ActionCalloutStyle")
+public typealias SecondaryInputCalloutStyle = ActionCalloutStyle