+++ Sources/KeyboardKit/_Deprecated/View+SecondaryInputCallout.swift
@@ -1,11 +1,3 @@
-//
-//  View+SecondaryInputCallout.swift
-//  KeyboardKit
-//
-//  Created by Daniel Saidi on 2021-01-06.
-//  Copyright © 2021 Daniel Saidi. All rights reserved.
-//
-
 import SwiftUI
 
 public extension View {
@@ -18,12 +10,10 @@ public extension View {
        - context: The context to bind against.
        - style: The style to apply to the view, by default `.standard`.
      */
+    @available(*, deprecated, renamed: "actionCallout")
     func secondaryInputCallout(
-        context: SecondaryInputCalloutContext,
+        context: ActionCalloutContext,
         style: SecondaryInputCalloutStyle = .standard) -> some View {
-        return ZStack {
-            self
-            SecondaryInputCallout(context: context, style: style)
-        }.coordinateSpace(name: SecondaryInputCalloutContext.coordinateSpace)
+        self.actionCallout(context: context, style: style)
     }
 }